refactor(monitor): split OpenLayers hook for CodeScene health
Break useMonitorOlMap into bootstrap, layer sync, and shared refs so delta
analysis stays above the code health threshold.
